requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysql-connector-python==9.5.0
packaging==25.0
Pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
django-anymail==13.1
reportlab==4.4.5
